Extract Event Debug Dimension  into separate class
