+++ php/src/class-settings.php
@@ -44,13 +44,6 @@ class Settings {
 	const HARDCODED_TRACKER_CONFIG_FEATURE = 'site_performance_tracker_vitals';
 
 	/**
-	 * Option eventDebug name
-	 *
-	 * @var string
-	 */
-	const OPTION_EVENT_DEBUG = 'eventDebug';
-
-	/**
 	 * Option web_vitals_tracking_ratio name
 	 *
 	 * @var string
@@ -73,6 +66,7 @@ class Settings {
 			new AnalyticsIdField(),
 			new MeasurementVersionDimensionField(),
 			new EventMetaDimensionField(),
+			new EventDebugDimensionField(),
 		);
 	}
 
@@ -138,14 +132,6 @@ class Settings {
 		}
 
 		add_settings_field(
-			'event_debug_dimension',
-			__( 'Event Debug Dimension', 'site-performance-tracker' ),
-			array( $this, 'event_debug_dimension_render' ),
-			self::PAGE_ID,
-			self::SECTION_ID
-		);
-
-		add_settings_field(
 			self::OPTION_WEB_VITALS_TRACKING_RATIO,
 			__( 'Web Vitals Tracking Ratio', 'site-performance-tracker' ),
 			array( $this, 'web_vitals_tracking_ratio_render' ),
@@ -159,13 +145,6 @@ class Settings {
 	 */
 	public function settings_section_callback() {
 		echo esc_html( __( 'Update Site Performance Tracker settings', 'site-performance-tracker' ) );
-	}
-
-	/**
-	 * Render Event Debug Dimension form input.
-	 */
-	public function event_debug_dimension_render() {
-		$this->render_dimention_option( self::OPTION_EVENT_DEBUG, 'dimension3', 'event debug dimension' );
 	}
 
 	/**
@@ -242,7 +221,7 @@ class Settings {
 				AnalyticsIdField::OPTION_TAG_ID                                 => '',
 				MeasurementVersionDimensionField::OPTION_MEASUREMENT_VERSION    => '',
 				EventMetaDimensionField::OPTION_EVENT_META                      => '',
-				self::OPTION_EVENT_DEBUG                                        => '',
+				EventDebugDimensionField::OPTION_EVENT_DEBUG                    => '',
 				self::OPTION_WEB_VITALS_TRACKING_RATIO                          => Plugin::TRACKING_DEFAULT_CHANCE,
 			),
 			$options
