+++ tests/phpunit/class-test-settings.php
@@ -201,7 +201,7 @@ class Test_Settings extends WP_UnitTestCase {
 		$this->assertSame( $field_name, $field['id'] );
 		$this->assertSame( 'Event Debug Dimension', $field['title'] );
 
-		$this->assertSame( array( $this->settings, 'event_debug_dimension_render' ), $field['callback'] );
+		$this->assertSame( array( $this->settings->fields[4], 'event_debug_dimension_render' ), $field['callback'] );
 
 		$this->assertSame( array(), $field['args'] );
 	}
@@ -658,8 +658,10 @@ EOD;
 	}
 
 	public function test_event_debug_dimension_render__empty_options() {
+		$this->settings->settings_init();
+
 		ob_start();
-		$this->settings->event_debug_dimension_render();
+		$this->settings->fields[4]->event_debug_dimension_render();
 		$result = ob_get_contents();
 		ob_end_clean();
 
@@ -673,10 +675,12 @@ EOD;
 	}
 
 	public function test_event_debug_dimension_render__set_eventDebug() {
+		$this->settings->settings_init();
+
 		add_option( 'spt_settings', array( 'eventDebug' => 'dimension3' ) );
 
 		ob_start();
-		$this->settings->event_debug_dimension_render();
+		$this->settings->fields[4]->event_debug_dimension_render();
 		$result = ob_get_contents();
 		ob_end_clean();
 
@@ -693,8 +697,10 @@ EOD;
 		global $tracker_config;
 		$tracker_config['eventDebug'] = 'dimension33';
 
+		$this->settings->settings_init();
+
 		ob_start();
-		$this->settings->event_debug_dimension_render();
+		$this->settings->fields[4]->event_debug_dimension_render();
 		$result = ob_get_contents();
 		ob_end_clean();
